% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-airstrikes.R
\docType{data}
\name{airstrikes}
\alias{airstrikes}
\title{airstrikes}
\format{
A tibble with 3938 rows and 4 variables:
\describe{
\item{date}{Date (YYYY-MM-DD)}
\item{longitude}{Longitudes (decimal)}
\item{latitude}{Latitudees (decimal)}
\item{type}{Types of airstrikes (airstrikes or shows of force (SOF))}
}
}
\usage{
airstrikes
}
\description{
A subset of airstrikes data in Iraq (March to June 2007)
}
\examples{
airstrikes
}
\keyword{datasets}
