% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-airstrikes-2006.R
\docType{data}
\name{airstrikes_2006}
\alias{airstrikes_2006}
\title{airstrikes_2006}
\format{
A tibble with 2101 rows and 4 variables:
\describe{
\item{date}{Date (YYYY-MM-DD)}
\item{longitude}{Longitudes (decimal)}
\item{latitude}{Latitudees (decimal)}
\item{type}{Types of airstrikes (airstrikes or shows of force (SOF))}
}
}
\usage{
airstrikes_2006
}
\description{
A dataset of airstrikes in Iraq in 2006 (2006/1/1-2006/9/24)
}
\examples{
airstrikes_2006
}
\keyword{datasets}
