% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-airstrikes_base.R
\docType{data}
\name{airstrikes_base}
\alias{airstrikes_base}
\title{airstrikes_base}
\format{
A tibble with 148 rows and 3 variables:
\describe{
\item{date}{Date}
\item{longitude}{Longitudes (decimal)}
\item{latitude}{Latitudees (decimal)}
}
}
\usage{
airstrikes_base
}
\description{
A subset of airstrikes data in Iraq (a subset of airstrikes in January 2006) that can be used to construct baseline densities
}
\examples{
airstrikes_base
}
\keyword{datasets}
