% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_owin_into_sf.R
\name{convert_owin_into_sf}
\alias{convert_owin_into_sf}
\title{Function: convert_owin_into_sf}
\usage{
convert_owin_into_sf(window)
}
\arguments{
\item{window}{An owin object}
}
\value{
A list of polygon, dataframe, sfc_POLYGON, sf, and SpatialPolygonsDataFrame objects
}
\description{
A function that takes an owin object and returns sf-related objects
}
