% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est.R
\name{geocausal.plot.est}
\alias{geocausal.plot.est}
\title{Plot estimates}
\usage{
geocausal.plot.est(estimates)
}
\arguments{
\item{estimates}{estimates obtained by `get_estimates` function}
}
\description{
Plot estimates
}
