% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{geocausal.plot.obs}
\alias{geocausal.plot.obs}
\title{Plot observed densities}
\usage{
geocausal.plot.obs(
  dens_1,
  dens_2 = NA,
  dens_3 = NA,
  actual_data = NA,
  time_unit = NA
)
}
\arguments{
\item{dens_1}{density 1}

\item{dens_2}{density 2 (if any). By default, `NA`.}

\item{dens_3}{density 3 (if any). By default, `NA`.}

\item{actual_data}{actual data in the form of `hyperframe$column`.}

\item{time_unit}{x-axis label of the output}
}
\description{
Plot observed densities
}
