% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_sum.R
\name{get_base_sum}
\alias{get_base_sum}
\title{Summarize baseline densities}
\usage{
get_base_sum(data, time_column)
}
\arguments{
\item{data}{dataframe}

\item{time_column}{the name of time variable}
}
\value{
list of the following:
    * `plot`: ggplot object that shows temporal distribution of data
    * `quantile`: quantiles of counts 
    * `mean`: the mean
    * `variance`: the variance
}
\description{
`get_base_sum()` takes dataframe and summarizes it.
}
\examples{
get_base_sum(data = airstrikes_base, 
             time_column = "date")
}
