% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_density.R
\name{get_baseline_density}
\alias{get_baseline_density}
\title{Function: get_baseline_density}
\usage{
get_baseline_density(
  data,
  coordinates = c("longitude", "latitude"),
  window,
  grayscale = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{coordinates}{A vector of column names for longitudes and latitudes (in this order)}

\item{window}{The window object of interest}

\item{grayscale}{Whether to return a plot in grayscale (by default, FALSE)}
}
\value{
A list of density (im object), its plot (ggplot), and a plot of original data (points; ggplot)
}
\description{
A function that takes a dataframe and returns the baseline densities using Scott's rule of thumb
}
