% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_summary.R
\name{get_baseline_summary}
\alias{get_baseline_summary}
\title{Function: get_baseline_summary}
\usage{
get_baseline_summary(data, time_column)
}
\arguments{
\item{data}{A dataframe}

\item{time_column}{The name of the column for time variable}
}
\value{
A list of a plot (ggplot) that shows temporal distribution of data, and the mean, variance, and quantiles of data
}
\description{
A function that takes a column of dataframe and generates a plot and a summary
}
