% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causal_contrasts.R
\name{get_causal_contrasts}
\alias{get_causal_contrasts}
\title{Function: get_causal_contrasts}
\usage{
get_causal_contrasts(
  estimates_scenario_1,
  estimates_scenario_2,
  grayscale,
  expectation_use_raw
)
}
\arguments{
\item{estimates_scenario_1}{A counterfactual scenario (the output of get_estimates function)}

\item{estimates_scenario_2}{Another counterfactual scenario}

\item{grayscale}{Whether to grayscale}

\item{expectation_use_raw}{If TRUE, the plot displays the raw expectations (if FALSE, max is set to 1)}
}
\value{
A list of ggplot objects and the data
}
\description{
A function that returns a figure with causal contrasts
}
