% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counterfactual_density.R
\name{get_counterfactual_density}
\alias{get_counterfactual_density}
\title{Function: get_counterfactual_density}
\usage{
get_counterfactual_density(
  expected_number,
  baseline_density,
  power_density = NA,
  window,
  grayscale = FALSE
)
}
\arguments{
\item{expected_number}{The expected number of observations.}

\item{baseline_density}{The baseline density (an im object)}

\item{power_density}{The power density (an im object)}

\item{window}{An owin object}
}
\description{
A function that takes the target number, baseline density, and power density,
and generates a hyperframe with point patterns
}
