% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counterfactual_sum_log_density.R
\name{get_counterfactual_sum_log_intensity}
\alias{get_counterfactual_sum_log_intensity}
\title{Function: get_counterfactual_sum_log_intensity}
\usage{
get_counterfactual_sum_log_intensity(counterfactual_density, treatment_data)
}
\arguments{
\item{counterfactual_density}{A counterfactual density (an im object)}

\item{treatment_data}{In the form of hyperframe$column}
}
\description{
A function that takes a hyperframe and returns the log counterfactual densities
ie, the numerator of the equation
}
