% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_based_exp.R
\name{get_dist_based_exp}
\alias{get_dist_based_exp}
\title{Get the expectation of treatment events with arbitrary distances}
\usage{
get_dist_based_exp(
  cf_sim_results,
  entire_window,
  dist_map,
  dist_map_unit = "km",
  grayscale = FALSE,
  use_raw = FALSE
)
}
\arguments{
\item{cf_sim_results}{output of `sim_cf_dens()`}

\item{entire_window}{owin object of the entire region}

\item{dist_map}{im object whose cell values are the distance from a focus (e.g., city)}

\item{dist_map_unit}{either `"km"` or `"mile"`}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}

\item{use_raw}{logical. `use_raw` specifies whether to use the raw value of expectations or percentiles.
By default, `FALSE`.}
}
\value{
ggplot object that summarizes how expectations change over distances from a focus
}
\description{
`get_dist_based_exp()` takes counterfactual densities and
and returns the expected number of treatment events based on distances
from a user-specified focus.
}
\examples{
# Baseline density
baseline <- get_base_dens(data = airstrikes_base, 
                          coordinates = c("longitude", "latitude"),
                          window = iraq_window,
                          grayscale = FALSE)
                          
# Density 1: Distance from Mosul
dist_from_mosul <- get_dist_focus(window = iraq_window,
                                  longitude = c(43.158),
                                  latitude = c(36.349),
                                  resolution = 0.5,
                                  grayscale = FALSE,
                                  mile = FALSE,
                                  preprocess = FALSE)

# Density 2: Distance from Baghdad
dist_from_baghd <- get_dist_focus(window = iraq_window,
                                  longitude = c(44.366),
                                  latitude = c(33.315),
                                  resolution = 0.5,
                                  grayscale = FALSE,
                                  mile = FALSE,
                                  preprocess = FALSE)
                                  
# Simulation of power density
sim_power_mosul <- sim_power_dens(target_dens = list(dist_from_baghd$distance_im),
                                  dens_manip = dist_from_mosul$distance_im,
                                  priorities = 1,
                                  priorities_manip = c(1, 2, 5, 10, 15, 50),
                                  window = iraq_window,
                                  grayscale = FALSE)

# Simulation of counterfactual density
sim_cf_mosul <- sim_cf_dens(expected_number = 3,
                            base_dens = baseline$density,
                            power_sim_results = sim_power_mosul,
                            window = iraq_window,
                            grayscale = FALSE)

# Distance based expectation             
get_dist_based_exp(cf_sim_results = sim_cf_mosul, 
                   entire_window = iraq_window, 
                   dist_map = dist_from_mosul$distance_im,
                   dist_map_unit = "km",
                   grayscale = FALSE,
                   use_raw = TRUE)
}
