% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_based_expectation.R
\name{get_distance_based_expectation}
\alias{get_distance_based_expectation}
\title{Function: get_distance_based_expectation}
\usage{
get_distance_based_expectation(
  counterfactual_simulation_results,
  entire_window,
  density_of_interest,
  distance_map,
  distance_map_unit = "km",
  grayscale = FALSE,
  expectation_use_raw = FALSE
)
}
\arguments{
\item{counterfactual_simulation_results}{An output of the simulate_counterfactual_density function}

\item{entire_window}{An owin object (the entire region)}

\item{density_of_interest}{A density that is being manipulated}

\item{distance_map}{An image object whose cell values are the distance from the focus (e.g., city)}

\item{distance_map_unit}{km or mile}

\item{grayscale}{Whether to use grayscale}

\item{expectation_use_raw}{Whether to use the actual expectation or proportion for the y-axis. By default, FALSE}
}
\value{
A ggplot object that summarizes how expectations change over distances from the focus
}
\description{
A function that takes a list of power densities
and returns simulated counterfactual densities
}
