% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_em_vec.R
\name{get_em_vec}
\alias{get_em_vec}
\title{convert a list of im objects to a vector}
\usage{
get_em_vec(
  em,
  outcome_pixel_count = NULL,
  time_after = TRUE,
  entire_window = NULL,
  lag
)
}
\arguments{
\item{em}{column of a hyperframe that summarizes effect modifier data. In the form of `hyperframe$column`.}

\item{outcome_pixel_count}{A list of integer-valued pixel images giving outcome event counts per spatial pixel,
typically obtained from `pixel_count_ppp()`.}

\item{time_after}{whether to include one unit time difference between treatment and outcome. By default = TRUE}

\item{entire_window}{owin object (the entire region of interest). If given, then the values outside the region will be set to `NA`.}

\item{lag}{integer that specifies lags to calculate causal estimates}
}
\value{
a numeric vector giving the effect modifier values, flattened across
pixels and time periods, with values outside `entire_window` (if supplied) set
to `NA`.
}
\description{
`get_em_vec()` get the vector form of a column of a hyperframe that summarizes the effect modifier data in heterogeneity analysis
}
\details{
The function `get_em_vec()` get the vector form of the effect modifier in the heterogeneity analysis. It is useful
if you want to construct the variance matrix `E_mat` that is passed to the function `get_cate()`
}
\references{
Zhou, L., Imai, K., Lyall, J. and Papadogeorgou, G. (2024). Estimating heterogeneous treatment effects for spatio-temporal causal inference: how economic assistance moderates the effects of airstrikes on insurgent violence. arXiv preprint. \doi{10.48550/arXiv.2412.15128}

Mukaigawara, M., Imai, K., Lyall, J. and Papadogeorgou, G. (2025). Spatiotemporal causal inference with arbitrary spillover and carryover effects. arXiv preprint. \doi{10.48550/arXiv.2504.03464}
}
\seealso{
Other data preparation functions:
\code{\link[=get_hfr]{get_hfr()}},
\code{\link[=get_hist]{get_hist()}},
\code{\link[=imls_to_arr]{imls_to_arr()}},
\code{\link[=pixel_count_ppp]{pixel_count_ppp()}},
\code{\link[=smooth_ppp]{smooth_ppp()}}
}
\concept{data preparation functions}
