% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history.R
\name{get_history}
\alias{get_history}
\title{Function: get_history}
\usage{
get_history(tt, Xt, Yt = NA, lag, window, x_only = TRUE)
}
\arguments{
\item{tt}{Point of time of interest}

\item{Xt}{A treatment column of a hyperframe}

\item{Yt}{An outcome column of a hyperframe}

\item{lag}{Lag to aggregate treatment and outcome}

\item{window}{Window, should be saved as an owin object}

\item{x_only}{Whether to generate treatment history only; by default, FALSE}
}
\value{
A list of histories
}
\description{
A function that takes hyperframe that was generated with get_hfr
and generates treatment and outcome histories
}
