% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_dens.R
\name{get_obs_dens}
\alias{get_obs_dens}
\title{Generate observed densities}
\usage{
get_obs_dens(hfr, dep_var, indep_var, ndim = 128, resolution = NULL, window)
}
\arguments{
\item{hfr}{hyperframe}

\item{dep_var}{The name of the dependent variable.
Since we need to obtain the observed density of treatment events,
`dep_var` should be the name of the treatment variable.}

\item{indep_var}{vector of names of independent variables (covariates)}

\item{ndim}{the number of grid cells that is used to generate observed densities.
By default = 128 (128 x 128). Notice that as you increase `ndim`, the process gets computationally demanding.}

\item{resolution}{the resolution in km per pixel. If specified, overrides `ndim`.
For example, `resolution = 5` creates ~5km x 5km grid cells.}

\item{window}{owin object}
}
\value{
list of the following:
     * `indep_var`: independent variables
     * `coef`: coefficients
     * `deviance`: deviance
     * `null_deviance`: null deviance
     * `dispersion`: dispersion parameter
     * `res_df`: average residuals as a dataframe
     * `intens_grid_cells`: im object of observed densities for each time period
     * `estimated_counts`: the number of events that is estimated by the poisson point process model for each time period
     * `sum_log_intens`: the sum of log intensities for each time period
     * `actual_counts`: the number of events (actual counts)
     * `window`: window object used as an input
}
\description{
`get_obs_dens()` takes a hyperframe and returns observed densities.
The output is used as propensity scores.
}
\details{
`get_obs_dens()` assumes the poisson point process model and
calculates observed densities for each time period. It depends on `spatstat.model::mppm()`.
Users should note that the coefficients in the output are not directly interpretable,
since they are the coefficients inside the exponential of the poisson model.
}
\examples{
\donttest{
# Prepare data: airstrikes (treatment) and insurgencies (outcome), Iraq 2006
dat <- rbind(airstrikes_2006[airstrikes_2006$type == "Airstrike", ],
             insurgencies_2006)
dat$type <- ifelse(dat$type == "Airstrike", "airstrike", "insurgency")
dat$time <- as.numeric(dat$date - min(dat$date) + 1)
dat <- dat[dat$time <= 60, ]
hfr <- get_hfr(data = dat, col = "type", window = iraq_window,
               time_col = "time", time_range = c(1, 60),
               coordinates = c("longitude", "latitude"), combine = FALSE)

# Covariate surface: distance from Baghdad
hfr$dist_bag <- rep(list(get_dist_focus(window = iraq_window, lon = 44.366,
                                        lat = 33.315, ndim = 64)), nrow(hfr))

# Observed density of the treatment (propensity score)
obs <- get_obs_dens(hfr, dep_var = "airstrike", indep_var = "dist_bag",
                    ndim = 64, window = iraq_window)
}
}
\references{
Papadogeorgou, G., Imai, K., Lyall, J. and Li, F. (2022). Causal inference with spatio-temporal data: estimating the effects of airstrikes on insurgent violence in Iraq. \emph{Journal of the Royal Statistical Society Series B}, 84(5), 1969--1999. \doi{10.1111/rssb.12548}

Mukaigawara, M., Imai, K., Lyall, J. and Papadogeorgou, G. (2025). Spatiotemporal causal inference with arbitrary spillover and carryover effects. arXiv preprint. \doi{10.48550/arXiv.2504.03464}
}
\seealso{
[get_cf_dens()], [get_est()]

Other density estimation functions:
\code{\link[=get_adaptive_baseline_dens]{get_adaptive_baseline_dens()}},
\code{\link[=get_base_dens]{get_base_dens()}},
\code{\link[=get_cf_dens]{get_cf_dens()}},
\code{\link[=get_cf_dens_adaptive]{get_cf_dens_adaptive()}},
\code{\link[=get_cf_sum_log_intens]{get_cf_sum_log_intens()}},
\code{\link[=get_power_dens]{get_power_dens()}},
\code{\link[=sim_cf_dens]{sim_cf_dens()}},
\code{\link[=sim_power_dens]{sim_power_dens()}}
}
\concept{density estimation functions}
