% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_density.R
\name{get_obs_density}
\alias{get_obs_density}
\title{Function: get_obs_density}
\usage{
get_obs_density(hfr, dep_var, indep_var, ngrid = 100, window)
}
\arguments{
\item{hfr}{A hyperframe}

\item{dep_var}{A dependent variable}

\item{indep_var}{A vector of names of independent variables}

\item{ngrid}{The number of grid cells. By default = 100}

\item{window}{The window object of interest}
}
\description{
A function that takes a hyperframe and returns the observed densities
ie, propensity scores; the denominators of the equation
}
