% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_density.R
\name{get_power_density}
\alias{get_power_density}
\title{Function: get_power_density}
\usage{
get_power_density(target_densities, priorities, window, grayscale = FALSE)
}
\arguments{
\item{target_densities}{A list of target densities}

\item{priorities}{A vector of priorities for each of target densities}

\item{window}{An owin object}
}
\description{
A function that takes the target densities and their priorities
and returns a power density image
}
