% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoothed_outcome.R
\name{get_smoothed_outcome}
\alias{get_smoothed_outcome}
\title{Function: get_smoothed_outcome}
\usage{
get_smoothed_outcome(
  data_interest,
  method,
  initialization = TRUE,
  sampling = 0.05
)
}
\arguments{
\item{data_interest}{Data to convert; should be in the form of "hyperframe$column"}

\item{method}{Methods for smoothing. Either "mclust" (fixed) or "abramson" (adaptive)}

\item{initialization}{Whether to use smaller samples to initialize mclust. Need to set seed for reproduction. By default = TRUE}

\item{sampling}{Determines the proportion of data to use for initialization. By default = 0.05, ie using 5% of samples first}
}
\description{
A function that takes a column of hyperframes and generates a smoothed ppp
}
