% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_bound.R
\name{get_var_bound}
\alias{get_var_bound}
\title{Calculate variance upper bounds}
\usage{
get_var_bound(estimates)
}
\arguments{
\item{estimates}{an object returned from `get_est()` function}
}
\value{
list of variance upper bounds for each scenario (`bound_haj`) and causal contrasts (`bound_tau_haj`). 
Note that this function returns variance upper bounds for Hajek estimators
}
\description{
A function that calculates variance upper bounds
}
\details{
`get_var_bound()` is an internal function to `get_estimates()` function, 
performing the estimation analysis after `get_est()` function.
}
\references{
Papadogeorgou, G., Imai, K., Lyall, J. and Li, F. (2022). Causal inference with spatio-temporal data: estimating the effects of airstrikes on insurgent violence in Iraq. \emph{Journal of the Royal Statistical Society Series B}, 84(5), 1969--1999. \doi{10.1111/rssb.12548}

Mukaigawara, M., Imai, K., Lyall, J. and Papadogeorgou, G. (2025). Spatiotemporal causal inference with arbitrary spillover and carryover effects. arXiv preprint. \doi{10.48550/arXiv.2504.03464}
}
\seealso{
Other causal effect estimation functions:
\code{\link[=get_cate]{get_cate()}},
\code{\link[=get_est]{get_est()}},
\code{\link[=get_estimates]{get_estimates()}},
\code{\link[=get_weighted_surf]{get_weighted_surf()}}
}
\concept{causal effect estimation functions}
