% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-insurgencies-2006.R
\docType{data}
\name{insurgencies_2006}
\alias{insurgencies_2006}
\title{insurgencies_2006}
\format{
A tibble with 37701 rows and 4 variables:
\describe{
\item{date}{Date (YYYY-MM-DD)}
\item{longitude}{Longitudes (decimal)}
\item{latitude}{Latitudees (decimal)}
\item{type}{Types of insurgencies (improvised explosive devices (IED), small arms fire (SAF), or other)}
}
}
\usage{
insurgencies_2006
}
\description{
A dataset of insurgencies data in Iraq in 2006 (2006/1/1-2006/9/24)
}
\examples{
insurgencies_2006
}
\keyword{datasets}
