% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-iraq_window.R
\docType{data}
\name{iraq_window}
\alias{iraq_window}
\title{iraq_window}
\format{
A polygonal object:
\describe{
\item{type}{Polygonal}
\item{xrange}{Range (longitude)}
\item{yrange}{Range (latitude)}
\item{bdry}{Boundaries}
\item{units}{Units}
}
}
\usage{
iraq_window
}
\description{
An owin object of Iraq
}
\examples{
iraq_window
}
\keyword{datasets}
