% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_sum_log_intens.R
\name{med_sum_log_intens}
\alias{med_sum_log_intens}
\title{Calculate the sum of conditional log densities of mediators}
\usage{
med_sum_log_intens(data, mediator, indep_var, time_var, model)
}
\arguments{
\item{data}{A dataframe (the treatment data)}

\item{mediator}{Name of the mediator}

\item{indep_var}{Independent variables to run regression analyses}

\item{time_var}{Name of the time variable}

\item{model}{Model selection. The current version supports Poisson only}
}
\value{
A numeric vector of sums of conditional log densities of mediators for each time period
}
\description{
A function that takes a dataframe and returns the sum of conditional log densities
of the mediators, which will be used for the numerator of the equation
}
