% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cflist.R
\name{plot.cflist}
\alias{plot.cflist}
\title{Plot simulated counterfactual densities}
\usage{
\method{plot}{cflist}(
  x,
  ...,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  grayscale = FALSE
)
}
\arguments{
\item{x}{an object of class `cflist`, typically the output of [sim_cf_dens()].}

\item{...}{additional arguments. Currently ignored.}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
A `ggplot` object (a faceted arrangement of counterfactual density
images, one per power).
}
\description{
Plot method for objects of class `cflist`, returned by
[sim_cf_dens()]. Takes the simulated counterfactual densities and their
powers and returns a faceted counterfactual density image over the range of
parameters.
}
\seealso{
[sim_cf_dens()]
}
