% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.distlist.R
\name{plot.distlist}
\alias{plot.distlist}
\title{Plot distance-based expectations}
\usage{
\method{plot}{distlist}(
  x,
  ...,
  dist_map_unit = "km",
  grayscale = FALSE,
  win_plot = FALSE,
  use_raw = FALSE
)
}
\arguments{
\item{x}{an object of class `distlist`, typically the output of [get_distexp()].}

\item{...}{additional arguments. Currently ignored.}

\item{dist_map_unit}{either `"km"` or `"mile"`}

\item{grayscale}{grayscale or not. By default, FALSE.}

\item{win_plot}{whether to plot windows as well. By default, FALSE}

\item{use_raw}{logical. `use_raw` specifies whether to use the raw value of expectations or percentiles.
By default, `FALSE`.}
}
\value{
A `ggplot` object. If `win_plot = TRUE`, the arranged window plots;
otherwise the expectation-versus-distance plot.
}
\description{
Plot method for objects of class `distlist`, returned by
[get_distexp()]. Visualizes the expected number (or proportion) of treatment
events covered as a function of distance from the focus, and optionally the
windows defined by distance quantiles.
}
\seealso{
[get_distexp()]
}
