% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hyperframe.R
\name{plot.hyperframe}
\alias{plot.hyperframe}
\title{Plot a hyperframe}
\usage{
\method{plot}{hyperframe}(
  x,
  ...,
  col,
  time_col = "time",
  range,
  lim = NA,
  main = "Image object",
  scalename = NA,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  combined = TRUE
)
}
\arguments{
\item{x}{an object of class `hyperframe`, typically the output of [get_hfr()].}

\item{...}{additional arguments. Currently ignored.}

\item{col}{the name/s of a column of interest.}

\item{time_col}{The name of the column of time variable. By default, `"time"`. Note that the time variable must be integers.}

\item{range}{vector that specifies the range of tiem variable (e.g., `c("2007-01-01", "2007-01-31")`)}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}

\item{main}{title
To specify multiple columns, users should list column names as a character vector.}

\item{scalename}{the name of the scale (for images only)}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{combined}{logical. `combined` specifies whether to combine all the point processes to one plot.
This argument applies only to the case when users specify one column with multiple time periods.
By default = TRUE}
}
\value{
A `ggplot` object visualizing the selected column(s) and time period(s).
}
\description{
Plot method for objects of class `hyperframe` (as created by
[get_hfr()]). Visualizes the selected point pattern (`ppp`) or pixel image
(`im`) column(s) over the specified time period(s), arranging multiple time
periods and/or columns by faceting or combining them.
}
\seealso{
[get_hfr()]
}
