% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.imlist.R
\name{plot.imlist}
\alias{plot.imlist}
\title{Plot a list of pixel images}
\usage{
\method{plot}{imlist}(
  x,
  ...,
  main = "image",
  lim = NA,
  transf = NULL,
  frame = 1,
  scalename = "Density",
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  grayscale = FALSE,
  ncol = NA,
  nrow = NA
)
}
\arguments{
\item{x}{an object of class `imlist` (a list of spatstat pixel images), for
example a sequence of density images produced by a `get_*_dens()` function.}

\item{...}{additional arguments. Currently ignored.}

\item{main}{title}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}

\item{transf}{a function to transform the pixel values (by default, NULL)}

\item{frame}{the element number of the list object (by default, 1)}

\item{scalename}{the name of the scale}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{grayscale}{grayscale or not. By default, FALSE.}

\item{ncol}{the number of columns (if plotitng multiple images at once)}

\item{nrow}{the number of rows (if plotting multiple images at once)}
}
\value{
A `ggplot` object. For a single frame, the rendered image; for
multiple frames, an arranged grid of images with a common legend.
}
\description{
Plot method for objects of class `imlist` (a list of spatstat
pixel images, such as a sequence of density images over time periods). Renders
one or more images as filled rasters over their observation window, arranging
multiple frames into a common-legend grid.
}
