% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.powerlist.R
\name{plot.powerlist}
\alias{plot.powerlist}
\title{Plot simulated power densities}
\usage{
\method{plot}{powerlist}(
  x,
  ...,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  grayscale = FALSE
)
}
\arguments{
\item{x}{an object of class `powerlist`, typically the output of [sim_power_dens()].}

\item{...}{additional arguments. Currently ignored.}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
A `ggplot` object (a faceted arrangement of power density images).
}
\description{
Plot method for objects of class `powerlist`, returned by
[sim_power_dens()]. Takes the simulated power densities and their priorities
and returns a faceted power density image over the range of parameters.
}
\seealso{
[sim_power_dens()]
}
