% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.obs}
\alias{plot.est}
\title{Plot results}
\usage{
plot(x, ...)

\method{plot}{obs}(x, dens_2 = NA, dens_3 = NA, actual_data = NA, time_unit = NA, ...)

\method{plot}{est}(x, ...)
}
\arguments{
\item{x}{an output object}

\item{...}{arguments}

\item{dens_2}{density 2 (if any). By default, `NA`.}

\item{dens_3}{density 3 (if any). By default, `NA`.}

\item{actual_data}{actual data in the form of `hyperframe$column`.}

\item{time_unit}{x-axis label of the output}
}
\description{
`plot` functions take the output and plot it.
}
\details{
Currently, observed densities (class: obs) and estimates (class: est) are supported by this function.
}
