% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.unit_ps.R
\name{plot.unit_ps}
\alias{plot.unit_ps}
\title{Plot unit-level propensity scores and balance}
\usage{
\method{plot}{unit_ps}(x, type = c("overlap", "balance"), ...)
}
\arguments{
\item{x}{an object returned by [get_unit_ps()].}

\item{type}{plot type: `"overlap"` displays propensity-score distributions
by observed treatment, and `"balance"` displays absolute standardized
differences before and after weighting.}

\item{...}{additional arguments. Currently ignored.}
}
\value{
A `ggplot` object.
}
\description{
Plot unit-level propensity scores and balance
}
\seealso{
[get_unit_ps()]
}
