% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.weights.R
\name{plot.weights}
\alias{plot.weights}
\title{Plot weights}
\usage{
\method{plot}{weights}(x, ..., type_weights = "standardized", binwidth = NULL)
}
\arguments{
\item{x}{an object of class `weights`, derived from the output of [get_est()].}

\item{...}{additional arguments. Currently ignored.}

\item{type_weights}{the type of weights to plot.
- If 'plot_weights' is `standardized`, histogram of standardized weights will be generated.
- If 'plot_weights' is `unstandardized`, histogram of unstandardized weights will be generated.
Default is `standardized`.}

\item{binwidth}{bin width of the histogram. Default is NULL}
}
\value{
A `ggplot` object showing histograms of the weights, faceted by
intervention.
}
\description{
Plot method for objects of class `weights`, derived from the
estimation produced by [get_est()]. Displays histograms of the (standardized
or unstandardized) inverse-probability weights, faceted by intervention.
}
\seealso{
[get_est()]
}
