% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_obs_dens.R
\name{predict_obs_dens}
\alias{predict_obs_dens}
\title{Perform out-of-sample prediction}
\usage{
predict_obs_dens(hfr, ratio, dep_var, indep_var, ngrid = 100, window)
}
\arguments{
\item{hfr}{hyperframe}

\item{ratio}{numeric. ratio between training and test sets}

\item{dep_var}{dependent variables}

\item{indep_var}{independent variables}

\item{ngrid}{the number of grids. By default, `100`.}

\item{window}{owin object}
}
\value{
list of the following:
     * `indep_var`: independent variables
     * `coef`: coefficients
     * `intens_grid_cells`: im object of observed densities for each time period
     * `estimated_counts`: the number of events that is estimated by the poisson point process model for each time period
     * `sum_log_intens`: the sum of log intensities for each time period
     * `training_row_max`: the max row ID of the training set
}
\description{
`predict_obs_dens()` performs out-of-sample prediction
(separating data into training and test sets).
It assumes that training and test sets have the same window.
}
