% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_obs_density.R
\name{predict_obs_density}
\alias{predict_obs_density}
\title{Function: predict_obs_density}
\usage{
predict_obs_density(hfr, ratio, dep_var, indep_var, ngrid = 100, window)
}
\arguments{
\item{ratio}{The ratio between training and test sets}
}
\description{
A function that performs out-of-sample prediction (separating data into training and test sets)
(Note that this funciton assumes the same window between training and test sets)
}
