% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counterfactual_density.R
\name{simulate_counterfactual_density}
\alias{simulate_counterfactual_density}
\title{Function: simulate_counterfactual_density}
\usage{
simulate_counterfactual_density(
  expected_number,
  baseline_density,
  power_simulation_results,
  window,
  grayscale = FALSE
)
}
\arguments{
\item{expected_number}{The expected number of observations}

\item{baseline_density}{The baseline density (an im object)}

\item{power_simulation_results}{The results obtained by simulate_power_density function}

\item{window}{An owin object}

\item{gray_scale}{By default, FALSE}
}
\description{
A function that takes a list of power densities
and returns simulated counterfactual densities
}
