% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_power_density.R
\name{simulate_power_density}
\alias{simulate_power_density}
\title{Function: simulate_power_density}
\usage{
simulate_power_density(
  target_densities,
  density_to_manipulate,
  priorities,
  priorities_for_manipulation,
  window,
  grayscale = FALSE
)
}
\arguments{
\item{target_densities}{A list of target densities}

\item{density_to_manipulate}{One target density for which we manipulate the value of alpha}

\item{priorities}{A vector of priorities for each of target densities}

\item{priorities_for_manipulation}{A vector of priority values}

\item{window}{An owin object}

\item{grayscale}{By default, FALSE}
}
\description{
A function that takes the target densities and their priorities
and returns a power density image over a range of parameters
}
