% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_base.R
\name{sum_base}
\alias{sum_base}
\title{Summarize baseline densities}
\usage{
sum_base(data, time_column)
}
\arguments{
\item{data}{dataframe}

\item{time_column}{the name of time variable}
}
\value{
list of the following:
    * `plot`: ggplot object that shows temporal distribution of data
    * `quantile`: quantiles of counts 
    * `mean`: the mean
    * `variance`: the variance
}
\description{
`sum_base()` takes dataframe and summarizes it.
}
\examples{
sum_base(data = airstrikes_base, 
         time_column = "date")
}
