% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_est.R
\name{sum_est}
\alias{sum_est}
\title{Summarize estimated results}
\usage{
sum_est(estimates)
}
\arguments{
\item{estimates}{an object returned from `get_est()` function}
}
\value{
list of results (as a dataframe) and windows
}
\description{
A function that summarizes estimated results
}
