% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.obs}
\alias{summary.est}
\title{Summarize results}
\usage{
summary(x)

\method{summary}{obs}(x)

\method{summary}{est}(x)
}
\arguments{
\item{x}{an output object}
}
\description{
`summary` functions take the output and summarize it.
}
\details{
Currently, observed densities (class: obs) and estimates (class: est) are supported by this function.
}
