% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.unit_ps.R
\name{summary.unit_ps}
\alias{summary.unit_ps}
\title{Summarize a unit-level propensity-score model}
\usage{
\method{summary}{unit_ps}(object, ...)
}
\arguments{
\item{object}{an object returned by [get_unit_ps()].}

\item{...}{additional arguments. Currently ignored.}
}
\value{
A list containing the model coefficients, probability summary, and
balance table.
}
\description{
Summarize a unit-level propensity-score model
}
\seealso{
[get_unit_ps()]
}
