% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_est.R
\name{vis_est}
\alias{vis_est}
\title{Visualize estimated results}
\usage{
vis_est(estimates)
}
\arguments{
\item{estimates}{an object returned from `get_est()` function}
}
\value{
list of two plots: one for differences in average weighted surfaces and another for results
}
\description{
A function that visualizes estimated results
}
\details{
The plot for differences in average weighted surfaces is comparing scenario 2 to scenario 1 
(i.e., the baseline is scenario 1).
}
