% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_hfr.R
\name{vis_hfr}
\alias{vis_hfr}
\title{Function: vis_hfr}
\usage{
vis_hfr(
  hfr,
  subtype_column,
  time_column = "time",
  range,
  combined = TRUE,
  scale_max = 100
)
}
\arguments{
\item{hfr}{A hyperframe to convert}

\item{subtype_column}{A column of interest}

\item{combined}{Whether to combine all the point processes to one plot. By default = TRUE}

\item{scale_max}{Choose the maximum of the color scale (for visualizing images; by default = 100)}

\item{time_range}{The range of dates (e.g., c("2007-02-23", "2007-02-28"))}
}
\description{
A function that generates a plot of point processes
}
