% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_owin_into_sf.R
\name{conv_owin_into_sf}
\alias{conv_owin_into_sf}
\title{Convert windows into sf objects}
\usage{
conv_owin_into_sf(window)
}
\arguments{
\item{window}{owin object}
}
\value{
list of polygon, dataframe, sfc_POLYGON, sf, and SpatialPolygonsDataFrame objects
}
\description{
`conv_owin_into_sf` takes an owin object and converts it to sf-related objects. 
This function is mostly an internal function of other functions.
}
