% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_owin_into_sf.R
\name{convert_owin_into_sf}
\alias{convert_owin_into_sf}
\title{Function: convert_owin_into_sf}
\usage{
convert_owin_into_sf(window)
}
\arguments{
\item{window}{owin object}
}
\value{
list of polygon, dataframe, sfc_POLYGON, sf, and SpatialPolygonsDataFrame objects
}
\description{
`convert_owin_into_sf` takes an owin object and converts it to sf-related objects. 
This function is mostly an internal function of other functions.
}
