% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_density.R
\name{get_baseline_density}
\alias{get_baseline_density}
\title{Function: get_baseline_density}
\usage{
get_baseline_density(
  data,
  coordinates = c("longitude", "latitude"),
  window,
  grayscale = FALSE
)
}
\arguments{
\item{data}{dataframe}

\item{coordinates}{vector of column names of longitudes and latitudes (in this order)}

\item{window}{owin object}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
list of the following:
    * `density`: im object of baseline density
    * `density_plot`: ggplot object of baseline density
    * `point_plot`: ggplot object of raw data
}
\description{
`get_baseline_density()` takes a dataframe and 
returns the baseline densities using Scott's rule of thumb.
}
