% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_summary.R
\name{get_baseline_summary}
\alias{get_baseline_summary}
\title{Function: get_baseline_summary}
\usage{
get_baseline_summary(data, time_column)
}
\arguments{
\item{data}{dataframe}

\item{time_column}{the name of time variable}
}
\value{
list of the following:
    * `plot`: ggplot object that shows temporal distribution of data
    * `quantile`: quantiles of counts 
    * `mean`: the mean
    * `variance`: the variance
}
\description{
`get_baseline_summary()` takes dataframe and summarizes it.
}
