% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causal_cont.R
\name{get_causal_cont}
\alias{get_causal_cont}
\title{Generate a figure with causal contrasts}
\usage{
get_causal_cont(
  scenario_1,
  scenario_2,
  dist_map_unit = "km",
  grayscale,
  use_raw
)
}
\arguments{
\item{scenario_1}{A counterfactual scenario (the output of get_estimates function)}

\item{scenario_2}{Another counterfactual scenario}

\item{dist_map_unit}{either `"km"` or `"mile"`}

\item{grayscale}{Whether to grayscale}

\item{use_raw}{If TRUE, the plot displays the raw expectations (if FALSE, max is set to 1)}
}
\value{
A list of ggplot objects and the data
}
\description{
A function that returns a figure with causal contrasts
}
\details{
`get_causal_cont()` returns a figure with causal contrasts. 
A causal contrast is the difference between expected numbers of outcome events of two counterfactual scenarios. 
The baseline scenario is `scenario_1`.
}
