% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causal_contrasts.R
\name{get_causal_contrasts}
\alias{get_causal_contrasts}
\title{Function: get_causal_contrasts}
\usage{
get_causal_contrasts(
  estimates_scenario_1,
  estimates_scenario_2,
  grayscale,
  expectation_use_raw
)
}
\arguments{
\item{estimates_scenario_1}{A counterfactual scenario}

\item{estimates_scenario_2}{Another counterfactual scenario}
}
\description{
A function that returns a figure with causal contrasts
}
