% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cf_dens_adaptive.R
\name{get_cf_dens_adaptive}
\alias{get_cf_dens_adaptive}
\title{Get counterfactual densities}
\usage{
get_cf_dens_adaptive(baseline_den, scale_factor)
}
\arguments{
\item{baseline_den}{baseline density (obs object obtained from `get_adaptive_baseline_dens`)}

\item{scale_factor}{a positive number that scales the baseline intensity}
}
\value{
a list of the following:
     * `intens_grid_cells`: im object of counterfactural intensities for each time period
     * `estimated_counts`: the number of events that is estimated by the poisson point process model for each time period
     * `sum_log_intens`: the sum of log intensities for each time period
}
\description{
`get_cf_dens_adaptive` takes the scale_factor, baseline density to construct counterfactural intensities .
}
