% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cf_sum_log_intens.R
\name{get_cf_sum_log_intens}
\alias{get_cf_sum_log_intens}
\title{Calculate the log counterfactual densities}
\usage{
get_cf_sum_log_intens(cf_dens, treatment_data)
}
\arguments{
\item{cf_dens}{A counterfactual density (an im object)}

\item{treatment_data}{In the form of hyperframe$column}
}
\value{
A numeric vector of sums of log densities for each time period
}
\description{
A function that takes a hyperframe and returns the log counterfactual densities
ie, the numerator of the equation
}
