% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counterfactual_density.R
\name{get_counterfactual_density}
\alias{get_counterfactual_density}
\title{Function: get_counterfactual_density}
\usage{
get_counterfactual_density(
  expected_number,
  baseline_density,
  power_density = NA,
  window,
  grayscale = FALSE
)
}
\arguments{
\item{expected_number}{the expected number of observations.}

\item{baseline_density}{baseline density (im object)}

\item{power_density}{power density (im object)}

\item{window}{owin object}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
A list of counterfactual density (im object) and a ggplot object
}
\description{
`get_counterfactual_density` takes the target (expected) number, baseline density, 
and power density, and generates a hyperframe with counterfactual densities.
}
\details{
There are two ways of generating counterfactual densities. 
First, users can keep the locations of observations as they are and change the expected number of observations. 
In this case, users do not have to set `power_density` and simply modify `expected_number`. 
Alternatively, users can shift the locations as well. In this case, `power_density` must be specified. 
To obtain power densities, refer to [get_power_density()].
}
