% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_line.R
\name{get_dist_line}
\alias{get_dist_line}
\title{Function: get_dist_line}
\usage{
get_dist_line(
  window,
  path_to_shapefile,
  line_data = NULL,
  grayscale,
  mile,
  resolution,
  preprocess = TRUE,
  ...
)
}
\arguments{
\item{window}{An owin object}

\item{path_to_shapefile}{Path to shapefile}

\item{line_data}{A sfc_MULTILINESTRING file (if available; if not the function creates it from a shapefile)}

\item{mile}{Whether to return the output in miles instead of kilometers}

\item{preprocess}{Whether to pick a roughly closest object first (by default, TRUE)}
}
\description{
A function that generates a distance map from lines (e.g., roads and rivers)
}
