% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_line.R
\name{get_dist_line}
\alias{get_dist_line}
\title{Get distance maps from lines and polygons}
\usage{
get_dist_line(
  window,
  path_to_shapefile,
  line_data = NULL,
  grayscale,
  mile = FALSE,
  resolution,
  preprocess = TRUE
)
}
\arguments{
\item{window}{owin object}

\item{path_to_shapefile}{path to shapefile}

\item{line_data}{sfc_MULTILINESTRING file (If available. If not, `get_dist_line()` creates it from a shapefile.)}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}

\item{mile}{logical. `mile` specifies whether to return the output in miles instead of kilometers (by default,  FALSE).}

\item{resolution}{resolution of raster objects}

\item{preprocess}{logical. `preprocess` specifies whether to first pick the potentially closest point.
It is recommended to set `preprocess = TRUE` if users need to obtain distances from many points.}
}
\value{
A list of im and ggplot object
}
\description{
`get_dist_line()` generates a distance map from lines and polygons.
}
