% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_based_expectation.R
\name{get_distance_based_expectation}
\alias{get_distance_based_expectation}
\title{Function: get_distance_based_expectation}
\usage{
get_distance_based_expectation(
  counterfactual_simulation_results,
  entire_window,
  density_of_interest,
  distance_map,
  distance_map_unit = "km",
  grayscale = FALSE,
  expectation_use_raw = FALSE
)
}
\arguments{
\item{counterfactual_simulation_results}{output of `simulate_counterfactual_density()`}

\item{entire_window}{owin object of the entire region}

\item{density_of_interest}{density to manipulate}

\item{distance_map}{im object whose cell values are the distance from a focus (e.g., city)}

\item{distance_map_unit}{`"km"` or `"mile"`}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}

\item{expectation_use_raw}{logical. `expectation_use_raw` specifies whether to use the raw value of expectations or percentiles. 
By default, `FALSE`.
whether to use the actual expectation or proportion for the y-axis. By default, `FALSE`.}
}
\value{
ggplot object that summarizes how expectations change over distances from a focus
}
\description{
`get_distance_based_expectation()` takes counterfactual densities and 
and returns the expected number of treatment events based on distances 
from a user-specified focus.
}
