% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev.R
\name{get_elev}
\alias{get_elev}
\title{Get elevation data}
\usage{
get_elev(load_path, ...)
}
\arguments{
\item{load_path}{path to the shp file (note: a folder)}

\item{...}{other parameters passed to `elevatr::get_elev_raster()`. The resolution argument z must be specified.}
}
\value{
an im object (unit: meters)
}
\description{
`get_elevation()` takes a directory that hosts shapefile and returns an owin object of altitudes.
}
