% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Function: get_elevation}
\usage{
get_elevation(load_path, ...)
}
\arguments{
\item{load_path}{A path to the shp file (note: a folder)}
}
\description{
A function that takes a directory that hosts shp file
and returns an owin object
}
