% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get elevation data}
\usage{
get_elevation(load_path, ...)
}
\arguments{
\item{load_path}{path to the shp file (note: a folder)}

\item{...}{other parameters passed to `elevatr::get_elev_raster()`}
}
\value{
list of a raster layer, an im object, and a ggplot object of altitudes (in meters).
}
\description{
`get_elevation()` takes a directory that hosts shapefile and returns an owin object of altitudes.
}
