% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Function: get_estimates}
\usage{
get_estimates(
  observed_density,
  counterfactual_density,
  treatment_data,
  smoothed_outcome,
  lag,
  entire_window,
  distance_map,
  distance_map_unit = "km",
  grayscale,
  expectation_use_raw
)
}
\arguments{
\item{observed_density}{observed density}

\item{counterfactual_density}{counterfactual density}

\item{treatment_data}{column of a hyperframe that summarizes treatment data. In the form of `hyperframe$column`.}

\item{smoothed_outcome}{column of a hyperframe that summarizes the smoothed outcome data}

\item{lag}{integer that specifies lags to calculate causal estimates}

\item{entire_window}{owin object (the entire region of interest)}

\item{distance_map}{im object (distance map)}

\item{distance_map_unit}{`"km"` or `"mile"`}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}

\item{expectation_use_raw}{logical. `expectation_use_raw` specifies whether to use the raw value of expectations or percentiles. 
By default, `FALSE`.}
}
\value{
list of the following:
`average_expected_events`: Hajek estimators (counts, the entire window)
`average_expected_events_quantiles`: percentiles of Hajek estimators
`weights`: weights
`average_weights`: mean of `weights`
`plot`: plot showing distance-based expectations
`distance_quantiles`: quantiles of distances
`expectation_plot`: plot of expectations
`window_plot`: plot of windows
}
\description{
`get_estimates()` generates causal estimates comparing two counterfactual scenarios.
}
