% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Function: get_estimates}
\usage{
get_estimates(
  observed_density,
  counterfactual_density,
  treatment_data,
  smoothed_outcome,
  lag,
  entire_window,
  distance_map,
  distance_map_unit = "km",
  grayscale,
  expectation_use_raw
)
}
\arguments{
\item{observed_density}{Observed density}

\item{counterfactual_density}{Counterfactual density}

\item{treatment_data}{A column of a hyperframe that summarizes treatment data}

\item{smoothed_outcome}{A column of a hyperframe that summarizes the smoothed outcome data}

\item{lag}{An integer (lags)}

\item{entire_window}{An owin object (the entire region of interest)}

\item{distance_map}{An im object}

\item{distance_map_unit}{"km" or "mile"}

\item{expectation_use_raw}{Whether to use the actual expectation or proportion for the y-axis. By default, FALSE}

\item{gray_scale}{Whether to use grayscale}
}
\value{
A list of an im object (weighted surface), counts of expected outcome events, weights, and plots
}
\description{
A function that takes a list of power densities
and returns simulated counterfactual densities
}
