% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Function: get_estimates}
\usage{
get_estimates(
  observed_density,
  counterfactual_density,
  treatment_data,
  smoothed_outcome,
  lag,
  entire_window,
  distance_map = dist_from_focus$distance_im,
  distance_map_unit = "km",
  grayscale,
  expectation_use_raw
)
}
\arguments{
\item{entire_window}{An owin object (the entire region)}

\item{expectation_use_raw}{Whether to use the actual expectation or proportion for the y-axis. By default, FALSE}

\item{counterfactual_simulation_results}{An output of the simulate_counterfactual_density function}

\item{density_of_interest}{A density that is being manipulated}

\item{gray_scale}{By default, FALSE}
}
\description{
A function that takes a list of power densities
and returns simulated counterfactual densities
}
