% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hfr.R
\name{get_hfr}
\alias{get_hfr}
\title{Function: get_hfr}
\usage{
get_hfr(
  data,
  subtype_column,
  window,
  time_column,
  time_range,
  coordinates = c("longitude", "latitude"),
  combined = TRUE
)
}
\arguments{
\item{data}{Data to convert}

\item{subtype_column}{Name of the column that specifies subtypes}

\item{window}{Window, which should be saved as an owin object}

\item{time_column}{Name of the column for time. By default, dates}

\item{time_range}{Range of time with min and max. By default, dates}

\item{coordinates}{Names of columns for coordinates. By default, x = longitude and y = latitude}

\item{combined}{Whether to generate output for all treatment and all outcomes combined. By default TRUE}
}
\value{
A hyperframe
}
\description{
A function that takes dataframes with coordinates
and generates a hyperframe with point patterns
}
