% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history.R
\name{get_history}
\alias{get_history}
\title{Function: get_history}
\usage{
get_history(tt, Xt, Yt = NA, lag, window, x_only = TRUE)
}
\arguments{
\item{tt}{values of the time variable of interest for which `get_history()` generates histories}

\item{Xt}{the name of a treatment column}

\item{Yt}{the name of an outcome column}

\item{lag}{numeric. `lag` specifies the number of time periods over which 
`get_history()` aggregates treatment and outcome columns.}

\item{window}{owin object.}

\item{x_only}{logical. `x_only` specifies whether to generate only treatment history (no outcome history). By default, `FALSE`.}
}
\value{
list of treatment and outcome histories
}
\description{
`get_history()` takes a hyperframe and time and columns of interest, and generates histories of events of interest.
}
