% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_density.R
\name{get_power_density}
\alias{get_power_density}
\title{Function: get_power_density}
\usage{
get_power_density(target_densities, priorities, window, grayscale = FALSE)
}
\arguments{
\item{target_densities}{list of target densities}

\item{priorities}{vector of priorities for each of target densities}

\item{window}{owin object}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
list of an im object and a ggplot object of power densities
}
\description{
`get_power_density()` takes the target densities and their priorities
and returns a power density.
}
