% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_bound.R
\name{get_var_bound}
\alias{get_var_bound}
\title{Calculate variance upper bounds}
\usage{
get_var_bound(estimates)
}
\arguments{
\item{estimates}{an object returned from `get_est()` function}
}
\value{
list of variance upper bounds for each scenario (`bound_haj`) and causal contrasts (`bound_tau_haj`). 
Note that this function returns variance upper bounds for Hajek estimators
}
\description{
A function that calculates variance upper bounds
}
\details{
`get_var_bound()` is an internal function to `get_estimates()` function, 
performing the estimation analysis after `get_est()` function.
}
