% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window.R
\name{get_window}
\alias{get_window}
\title{Function: get_window}
\usage{
get_window(load_path)
}
\arguments{
\item{load_path}{A path to the shp file (note: a folder)}
}
\value{
An owin object
}
\description{
A function that takes a directory that hosts shp file
and returns an owin object
}
