% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window.R
\name{get_window}
\alias{get_window}
\title{Generate a window}
\usage{
get_window(load_path, target_crs = NULL, unit_scale = 1000)
}
\arguments{
\item{load_path}{path to the shp file}

\item{target_crs}{target CRS (if users want to specify it; as an EPSG code)}

\item{unit_scale}{parameter to convert meters to kilometers (by default = 1000 so that the unit is set to km)}
}
\value{
owin object
}
\description{
`get_window()` takes a directory that hosts a shapefile
and returns an owin object.
}
