% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window.R
\name{get_window}
\alias{get_window}
\title{Generate a window}
\usage{
get_window(load_path)
}
\arguments{
\item{load_path}{path to the shp file}
}
\value{
owin object
}
\description{
`get_window()` takes a directory that hosts a shapefile
and returns an owin object.
}
