% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-insurgencies.R
\docType{data}
\name{insurgencies}
\alias{insurgencies}
\title{insurgencies}
\format{
A tibble with 68573 rows and 4 variables:
\describe{
\item{date}{Date (YYYY-MM-DD)}
\item{longitude}{Longitudes (decimal)}
\item{latitude}{Latitudees (decimal)}
\item{type}{Types of insurgencies (improvised explosive devices (IED), small arms fire (SAF), or other)}
}
}
\usage{
insurgencies
}
\description{
A subset of insurgencies data in Iraq (March to June 2007)
}
\examples{
insurgencies
}
\keyword{datasets}
