% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cate.R
\name{plot.cate}
\alias{plot.cate}
\title{Plot estimated CATE}
\usage{
\method{plot}{cate}(
  x,
  ...,
  categorical = 0,
  scale = 1,
  xrange = NULL,
  main = "",
  xlab = "",
  ylim = NULL
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{categorical}{boolean or a vector of boolean values indicating whether each value in `eval_values` should be treated as a categorical (TRUE) or continuous (FALSE). Default is `FALSE`.}

\item{scale}{a positive number specifying the scale by which the estimates will be scaled. If provided, the estimates will be scaled by this value. Default is NULL, which means no scaling is applied.}

\item{xrange}{an optional vector of two values the range of x shown.}

\item{main}{title}

\item{xlab}{label of x-axis}

\item{ylim}{an optional vector of two values specifying the limits of y}
}
\description{
Plot estimated CATE
}
