% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dens.R
\name{plot.dens}
\alias{plot.dens}
\title{Plot densities}
\usage{
\method{plot}{dens}(x, ..., main = "Density", grayscale = "FALSE")
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{main}{title}

\item{grayscale}{whether to use grayscale. By default, FALSE.}
}
\description{
Plot densities
}
