% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dens.R
\name{plot.dens}
\alias{plot.dens}
\title{Plot densities}
\usage{
\method{plot}{dens}(
  x,
  ...,
  main = "Image object",
  scalename = "Density",
  grayscale = "FALSE",
  lim = NA
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{main}{title}

\item{scalename}{the name of the scale (for images only)}

\item{grayscale}{whether to use grayscale. By default, FALSE.}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}
}
\description{
Plot densities
}
