% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.distlist.R
\name{plot.distlist}
\alias{plot.distlist}
\title{Plot distance-based expectations}
\usage{
\method{plot}{distlist}(
  x,
  ...,
  dist_map_unit = "km",
  grayscale = FALSE,
  win_plot = FALSE,
  use_raw = FALSE
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{dist_map_unit}{either `"km"` or `"mile"`}

\item{grayscale}{grayscale or not. By default, FALSE.}

\item{win_plot}{whether to plot windows as well. By default, FALSE}

\item{use_raw}{logical. `use_raw` specifies whether to use the raw value of expectations or percentiles.
By default, `FALSE`.}
}
\description{
Plot distance-based expectations
}
