% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.est.R
\name{plot.est}
\alias{plot.est}
\title{Plot estimates}
\usage{
\method{plot}{est}(x, ..., surface = FALSE, lim = NA)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{surface}{whether to produce the surface plot. By default, FALSE}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector of max and min}
}
\description{
Plot estimates
}
