% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hyperframe.R
\name{plot.hyperframe}
\alias{plot.hyperframe}
\title{Plot estimates}
\usage{
\method{plot}{hyperframe}(x, ..., col, time_col = "time", range, scalename = NA, combined = TRUE)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{col}{the name/s of a column of interest.
To specify multiple columns, users should list column names as a character vector.}

\item{time_col}{The name of the column of time variable. By default, `"time"`. Note that the time variable must be integers.}

\item{range}{vector that specifies the range of tiem variable (e.g., `c("2007-01-01", "2007-01-31")`)}

\item{scalename}{the name of the scale (for images only)
This argument applies only to the case when users specify one column with multiple time periods.
By default = TRUE}

\item{combined}{logical. `combined` specifies whether to combine all the point processes to one plot.}
}
\description{
Plot estimates
}
