% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hyperframe.R
\name{plot.hyperframe}
\alias{plot.hyperframe}
\title{Plot estimates}
\usage{
\method{plot}{hyperframe}(
  x,
  ...,
  col,
  time_col = "time",
  range,
  lim = NA,
  scalename = NA,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  combined = TRUE
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{col}{the name/s of a column of interest.
To specify multiple columns, users should list column names as a character vector.}

\item{time_col}{The name of the column of time variable. By default, `"time"`. Note that the time variable must be integers.}

\item{range}{vector that specifies the range of tiem variable (e.g., `c("2007-01-01", "2007-01-31")`)}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}

\item{scalename}{the name of the scale (for images only)}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{combined}{logical. `combined` specifies whether to combine all the point processes to one plot.
This argument applies only to the case when users specify one column with multiple time periods.
By default = TRUE}
}
\description{
Plot estimates
}
