% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.im.R
\name{plot.im}
\alias{plot.im}
\title{Plot densities}
\usage{
\method{plot}{im}(x, ..., main = "Distance", window, grayscale = "FALSE")
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{main}{title}

\item{window}{an owin object}

\item{grayscale}{whether to use grayscale. By default, FALSE.}
}
\description{
Plot densities
}
