% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.im.R
\name{plot.im}
\alias{plot.im}
\title{Plot im}
\usage{
\method{plot}{im}(
  x,
  ...,
  main = "Image object",
  scalename = "Density",
  grayscale = "FALSE",
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  lim = NA
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{main}{title}

\item{scalename}{the name of the scale (for images only)}

\item{grayscale}{whether to use grayscale. By default, FALSE.}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}
}
\description{
Plot im
}
