% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.list.R
\name{plot.list}
\alias{plot.list}
\title{Plot lists}
\usage{
\method{plot}{list}(
  x,
  ...,
  main = "list",
  lim = NA,
  transf = NULL,
  frame = 1,
  combined = TRUE,
  scalename = "Density",
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  grayscale = FALSE,
  ncol = NA,
  nrow = NA
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{main}{title}

\item{lim}{limits of the scale. By default, NA. To set limits manually, provide a vector or max and min}

\item{transf}{a function to transform the pixel values (by default, NULL)
This argument applies only to the case when users specify one column with multiple time periods.
By default = TRUE}

\item{frame}{the element number of the list object (by default, 1)}

\item{combined}{logical. `combined` specifies whether to combine all the point processes to one plot.}

\item{scalename}{the name of the scale}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{grayscale}{grayscale or not. By default, FALSE.}

\item{ncol}{the number of columns (if plotitng multiple images at once)}

\item{nrow}{the number of rows (if plotting multiple images at once)}
}
\description{
Plot lists
}
