% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.obs.R
\name{plot.obs}
\alias{plot.obs}
\title{Plot observed densities}
\usage{
\method{plot}{obs}(
  x,
  ...,
  dens_2 = NA,
  dens_3 = NA,
  lim = c(-1, 1),
  window,
  time_unit = NA,
  combined = TRUE
)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{dens_2}{density 2 (if any). By default, `NA`.}

\item{dens_3}{density 3 (if any). By default, `NA`.}

\item{lim}{limits of the scale for the average residual fields. By default, c(0, -1). To set limits manually, provide a vector of max and min}

\item{window}{owin object to plot the average residual fields}

\item{time_unit}{x-axis label of the output}

\item{combined}{whether to combine the two plots. By default, TRUE. If TRUE,
then the plot function produces one ggplot object. If FALSE, three objects (two ggplot and one dataframe) will be produced.}
}
\description{
Plot observed densities
}
