% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.obs.R
\name{plot.obs}
\alias{plot.obs}
\title{Plot observed densities}
\usage{
\method{plot}{obs}(x, ..., dens_2 = NA, dens_3 = NA, actual_data = NA, time_unit = NA)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{dens_2}{density 2 (if any). By default, `NA`.}

\item{dens_3}{density 3 (if any). By default, `NA`.}

\item{actual_data}{actual data in the form of `hyperframe$column`.}

\item{time_unit}{x-axis label of the output}
}
\description{
Plot observed densities
}
