% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.powerlist.R
\name{plot.powerlist}
\alias{plot.powerlist}
\title{Plot simulated power densities}
\usage{
\method{plot}{powerlist}(
  x,
  ...,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF"),
  grayscale = FALSE
)
}
\arguments{
\item{x}{input (should be the output of the `sim_power_dens()` function)}

\item{...}{arguments passed on to the function}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}
}
\value{
list of densities, plot, and priorities
}
\description{
A function that takes the simulated power densities and their priorities
and returns a power density image over a range of parameters
}
