% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ppplist.R
\name{plot.ppplist}
\alias{plot.ppplist}
\title{Plot point patterns}
\usage{
\method{plot}{ppplist}(x, ..., frame = 1, main = "ppp", combined = TRUE)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{frame}{the element number of the list object (by default, 1)}

\item{main}{title}

\item{combined}{logical. `combined` specifies whether to combine all the point processes to one plot.
This argument applies only to the case when users specify one column with multiple time periods.
By default = TRUE}
}
\description{
Plot point patterns
}
