% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.supthin.R
\name{plot.supthin}
\alias{plot.supthin}
\title{Plot the results of superthinning tests}
\usage{
\method{plot}{supthin}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}
}
\description{
Plot the results of superthinning tests
}
