% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.weights.R
\name{plot.weights}
\alias{plot.weights}
\title{Plot weights}
\usage{
\method{plot}{weights}(x, ..., type_weights = "standardized", binwidth = NULL)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed on to the function}

\item{type_weights}{the type of weights to plot.
- If 'plot_weights' is `standardized`, histogram of standardized weights will be generated.
- If 'plot_weights' is `unstandardized`, histogram of unstandardized weights will be generated.
Default is `standardized`.}

\item{binwidth}{bin width of the histogram. Default is NULL}
}
\description{
Plot weights
}
