% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cate.R
\name{print.cate}
\alias{print.cate}
\title{Print results}
\usage{
\method{print}{cate}(x, ...)
}
\arguments{
\item{x}{an output object}

\item{...}{arguments passed on to the function}
}
\description{
`print` functions take the output and print the summary of it.
}
\details{
Currently, observed densities (class: obs), estimates (class: est) and heterogeneity estimates (class: cate) are supported by this function.
}
