% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.est.R
\name{print.est}
\alias{print.est}
\title{Print results}
\usage{
\method{print}{est}(x, ...)
}
\arguments{
\item{x}{an output object}

\item{...}{arguments passed on to the function}
}
\description{
`print` functions take the output and print the summary of it.
}
\details{
Currently, observed densities (class: obs) and estimates (class: est) are supported by this function.
}
