% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cf_dens.R
\name{sim_cf_dens}
\alias{sim_cf_dens}
\title{Simulate counterfactual densities}
\usage{
sim_cf_dens(
  expected_number,
  base_dens,
  power_sim_results,
  window,
  grayscale = FALSE,
  color = c("white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", "#771F59FF")
)
}
\arguments{
\item{expected_number}{the expected number of observations}

\item{base_dens}{the baseline density (im object)}

\item{power_sim_results}{the results obtained by `simulate_power_density()`}

\item{window}{owin object}

\item{grayscale}{logical. `grayscale` specifies whether to convert plot to grayscale (by default, FALSE).}

\item{color}{the color scale. By default, "white", "#F8DAC5FF", "#F4825AFF", "#D2204CFF", and "#771F59FF".}
}
\value{
list of counterfactual densities, a ggplot, and priorities
}
\description{
`sim_cf_dens()` takes a list of power densities
and returns simulated counterfactual densities.
}
