% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cf_dens.R
\name{sim_cf_dens}
\alias{sim_cf_dens}
\title{Simulate counterfactual densities}
\usage{
sim_cf_dens(expected_number, base_dens, power_sim_results, window)
}
\arguments{
\item{expected_number}{the expected number of observations}

\item{base_dens}{the baseline density (im object)}

\item{power_sim_results}{the results obtained by `simulate_power_density()`}

\item{window}{owin object}
}
\value{
list of counterfactual densities, power as numerics, and expected number as a numeric
}
\description{
`sim_cf_dens()` takes a list of power densities
and returns simulated counterfactual densities.
}
