% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_dens.R
\name{sim_power_dens}
\alias{sim_power_dens}
\title{Simulate power densities}
\usage{
sim_power_dens(target_dens, dens_manip, priorities, priorities_manip, window)
}
\arguments{
\item{target_dens}{list of target densities. This should always be a list, even if there is only one target density.}

\item{dens_manip}{a target density for which we manipulate the value of priorities}

\item{priorities}{numeric. `priorities` specifies the priority for the target density that we do not manipulate.}

\item{priorities_manip}{vector of priorities for the density that we manipulate.}

\item{window}{owin object}
}
\value{
list of densities and priorities
}
\description{
A function that takes the target densities and their priorities
and returns a power density image over a range of parameters
}
