% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.est.R
\name{summary.est}
\alias{summary.est}
\title{Summarize results}
\usage{
\method{summary}{est}(object, ...)
}
\arguments{
\item{object}{an output object}

\item{...}{arguments passed on to the function}
}
\description{
`summary` functions take the output and summarize it.
}
\details{
Currently, observed densities (class: obs) and estimates (class: est) are supported by this function.
}
