% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_obs_density.R
\name{vis_obs_density}
\alias{vis_obs_density}
\title{Function: vis_obs_density}
\usage{
vis_obs_density(
  actual_data,
  density_1,
  density_2 = NA,
  density_3 = NA,
  color_actual = "darkgrey",
  color_dens_1 = "#D55E00",
  color_dens_2 = "#0072B2",
  color_dens_3 = "#009E73",
  time_unit
)
}
\arguments{
\item{actual_data}{actual data in the form of `hyperframe$column`.}

\item{density_1}{im object that represents densities}

\item{density_2}{density 2 (if any). By default, `NA`.}

\item{density_3}{density 3 (if any). By default, `NA`.}

\item{color_actual}{color for actual data. By default, `"darkgrey"`.}

\item{color_dens_1}{color for actual data. By default, `"#D55E00"`.}

\item{color_dens_2}{color for actual data. By default, `"#0072B2"`.}

\item{color_dens_3}{color for actual data. By default, `"#009E73"`.}

\item{time_unit}{x-axis label of the output}
}
\value{
list of dataframe and two ggplot objects for comparison of densities and residuals
}
\description{
`vis_obs_density()` visualizes actual and predicted counts 
generated by [get_obs_density()].
}
