% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_obs_density.R
\name{vis_obs_density}
\alias{vis_obs_density}
\title{Function: vis_obs_density}
\usage{
vis_obs_density(
  actual_data,
  density_1,
  density_2 = NA,
  density_3 = NA,
  color_actual = "darkgrey",
  color_dens_1 = "#D55E00",
  color_dens_2 = "#0072B2",
  color_dens_3 = "#009E73",
  time_unit
)
}
\arguments{
\item{time_unit}{The label of the x-axis (time)}
}
\description{
A function that performs the visualization of actual counts and predicted counts
}
